MAINT: renamed time folder into time_stepper
