+++ tests/time_stepper/test_time_stepper.py
@@ -26,8 +26,8 @@ from porepy.numerics.nonlinear.nonlinear_solver_status import (
     NonlinearSolverStatusFailed,
 )
 from porepy.numerics.nonlinear.nonlinear_solvers import NewtonSolver
-from porepy.time.time_step_control import TimeManager
-from porepy.time.time_stepper import TimeStepper
+from porepy.time_stepper.time_step_control import TimeManager
+from porepy.time_stepper.time_stepper import TimeStepper
 from porepy.viz.solver_statistics import (
     NonlinearSolverAndTimeStatistics,
     SolverStatisticsFactory,
